move checkLockout logic to LeagueEntries component
